Move shouldTargetCheckpointV2 and buildCheckpointFromConfigOverride to StreamerCheckpointUtils
