+++ hudi-utilities/src/main/java/org/apache/hudi/utilities/streamer/StreamerCheckpointUtils.java
@@ -26,6 +26,8 @@ import org.apache.hudi.common.table.HoodieTableMetaClient;
 import org.apache.hudi.common.table.HoodieTableVersion;
 import org.apache.hudi.common.table.checkpoint.Checkpoint;
 import org.apache.hudi.common.table.checkpoint.CheckpointUtils;
+import org.apache.hudi.common.table.checkpoint.StreamerCheckpointV1;
+import org.apache.hudi.common.table.checkpoint.UnresolvedStreamerCheckpointBasedOnCfg;
 import org.apache.hudi.common.table.timeline.HoodieInstant;
 import org.apache.hudi.common.table.timeline.HoodieTimeline;
 import org.apache.hudi.common.table.timeline.TimelineUtils;
@@ -44,8 +46,8 @@ import lombok.extern.slf4j.Slf4j;
 
 import java.io.IOException;
 
+import static org.apache.hudi.common.table.checkpoint.CheckpointUtils.DATASOURCES_NOT_SUPPORTED_WITH_CKPT_V2;
 import static org.apache.hudi.common.table.checkpoint.CheckpointUtils.HOODIE_INCREMENTAL_SOURCES;
-import static org.apache.hudi.common.table.checkpoint.CheckpointUtils.buildCheckpointFromConfigOverride;
 import static org.apache.hudi.common.table.checkpoint.StreamerCheckpointV2.STREAMER_CHECKPOINT_KEY_V2;
 import static org.apache.hudi.common.table.checkpoint.StreamerCheckpointV2.STREAMER_CHECKPOINT_RESET_KEY_V2;
 import static org.apache.hudi.common.table.timeline.InstantComparison.LESSER_THAN;
@@ -55,6 +57,28 @@ import static org.apache.hudi.table.upgrade.UpgradeDowngrade.needsUpgradeOrDowng
 
 @Slf4j
 public class StreamerCheckpointUtils {
+
+  /**
+   * Wraps a user-supplied checkpoint override string. For HoodieIncrSource family on table version
+   * 8+, returns {@link UnresolvedStreamerCheckpointBasedOnCfg} so the source can resolve V1/V2
+   * cursor semantics from the override's prefix; for everything else returns V1.
+   */
+  public static Checkpoint buildCheckpointFromConfigOverride(
+      String sourceClassName, int writeTableVersion, String checkpointToResume) {
+    return shouldTargetCheckpointV2(writeTableVersion, sourceClassName)
+        ? new UnresolvedStreamerCheckpointBasedOnCfg(checkpointToResume)
+        : new StreamerCheckpointV1(checkpointToResume);
+  }
+
+  /**
+   * True only for the HoodieIncrSource family on table version 8+, where V2 (completion-time)
+   * cursor semantics differ from V1 (requested-time). Every other source operates on V1 only.
+   */
+  public static boolean shouldTargetCheckpointV2(int writeTableVersion, String sourceClassName) {
+    return writeTableVersion >= HoodieTableVersion.EIGHT.versionCode()
+        && HOODIE_INCREMENTAL_SOURCES.contains(sourceClassName)
+        && !DATASOURCES_NOT_SUPPORTED_WITH_CKPT_V2.contains(sourceClassName);
+  }
 
   /**
    * The first phase of checkpoint resolution - read the checkpoint configs from 2 sources and resolve
